requirements.txt
ultralytics==8.4.19
pyyaml>=6.0
